First prototype/idea of 0.2.0 - Docs not updated
