+++ lectio/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa
-from .lectio import Lectio, Module
+from .lectio import Lectio
+from . import user
 from . import lectio
 from . import exceptions
