+++ lectio/helpers/__init__.py
@@ -1,0 +1,1 @@
+from .schedule import *
